feat: Acquire lock before trying to export
